MS_TimeSeriesDBConnector/requirements.txt
cherrypy==18.8.0
paho-mqtt==2.2.1
influxdb-client==1.36.1
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0